\name{Check.pop}
\alias{Check.pop}
\title{
Check population file
}
\description{
Carries out a range of checks on the population file to detect the most commonly encountered errors. Provides a barchart showing the population structure.
}
\usage{
Check.pop(popfile, desvars)
}
\arguments{
  \item{popfile}{
   dataframe containing information for all plots in the population.
}
  \item{desvars}{
   vector containing the names of the design variables.
}
}
\value{
Reports on any errors found and produces a barchart.
}
\author{
G. Melville}
\seealso{
\code{\link{NC.sample}}.
}
\examples{
## Check.pop(popfile, desvars) 
}

